Return "file not found" error on FoE read
This change is crucial because it dictates the logic in the read_file function. When Motion Master detects that a file is not found, as opposed to encountering another type of error, it will attempt to read that file as a zip file in chunks. For more details, refer to Nikola's explanation here: https://github.com/synapticon/SOEM-old/commit/67434076c8d60d63ea665df91ed8b1f9745d35ee